+++ changelog.md
@@ -1,3 +1,8 @@
+##### 14.03.2019 (v.0.4.3)
+
+* Added UWP application support (based on commit by Mark Wiseman).
+* Added some common initialization variables to settings page configuration block.
+
 ##### 13.09.2018 (v.0.4.2)
 
 * Added APPEAR_OVER_TASKBAR custom initialization option, see the main site for details.